Make CLI completers stop filling values when command line is complete
Added <null> tags to terminate <completers> lists in the CLI shell
config.  These null values stop completers from adding extra
arguments at the end of commands.  If a <null> is not added
at the end of a list of completers, that means use the last
completer as many times as possible.