Aliased $OCX variables to the X number for onos and onos-log commands.
I.e. running "onos 1" is equivalent to running "onos $OC1".
"onos-log 2" is equivalent to "onos-log $OC2".

If the argument is not a number there should be no change to the behaviour
of the script (original invocations should still work as expected).

Will make this change for other commands in the future if there are no issues.

Change-Id: I7621cce9076c088d3bcb1aa4d6c8f8f8525823ca