fix for flows stuck in pending add state
Problem is due to two packet requests from different services
for the same selector resulting in the same flow rule.
Since these rules where the same, only one ended up on the data plane,
but sadly only the first request made it into the flow service meaning
that in cases where what was in the flow service does not match what is
on the dataplane resulting in a PENDING_ADD situation.

Change-Id: I4c03e753be6e198e04f0b5263a2aa8cf2edc51e1