ONOS-1926: Handle unsupported match field
Following is the main idea of this approach.

An UnsupportedOperationException indicates that:
    1. The OF version is not supported (other than OF10 and OF13)
    2. The flow contains unsupported match field
       (e.g. install ICMPv6 match to an OF10 switch)
I believe retrying are not going to help in both cases.
Therefore, I directly change the flow state into FAILED.

In case 2, if the switch is reconnected/reconfigured with correct
OF version, the flow will be reinstalled to the switch correctly.

Change-Id: I954f3597a77934e46695b82a6d17363d41636ebe