add Mac and hostname to POST url
This commit is to handle the case:
We receive a DELETE to delete a vBNG, and then receive
a POST with the same private IP address to create vBNG.

At this moment, if the old vCPE is not down, and the new
vCPE is not up yet, onos will use the host with old
vCPE mac address to create the intent, which is not
correct, and hard to solve. Since onos does not know
the host in the system is old vCPE or new vCPE.

Thus, we change to ask XOS to post us host MAC
and hostname together with the private IP address.

Change-Id: I011924aada589ea81d80a479aefa41753fa2e223