Override remote public key with local one
A problem is observed when running stc with same local and remote username.
The public key of local user is overridden by the key of remote user.
As a result, stc cannot login to the remote host.

For now, we let the public key of local user overrides the key of remote user.
But eventually we need a better solution for username collision.

Change-Id: I750fd129a2dcd4920683599b6c562d3beb84e95d