[ONOS-2885] Resolve flow installation failure bug
There is a time gap between
controller is elected as the master of a switch
and
controller actually becomes the master of a switch (receives ROLE_REPLY).
During this time gap, OF messages destined for the switch will be sent to this controller.
However, the OF message can only be sent to the switch after controller receives the ROLE_REPLY.

Change-Id: I32d68384226e272658a0c9de2d32ae9e1cc39b6a