Set -u and -e, and respect ONOS_HOME
If errors or undefined variables are introduced into this script,
we should exit with an error. In particular, we should exit with
an error if KARAF_VERSION is not defined rather than failing
semi-silently.

If ONOS_HOME is set, we should respect (and use) its value.

If ONOS_HOME is not set (as it will not be in the init or service
environment), intuit its value based on the path to this script.

Change-Id: I91c733fddcb2241a95466667e4b57ce0f74a8194