Fix for ONOS-4484 flows left behind after intent reroute
During the installation process, the list of flows to remove and
the list of flows to install were computed into the same variable,
so the list of adds overwrote the list of removes. The fix is to
put them all onto the same list.

Change-Id: I1f0c7f0a7b3c76f50afdb121dbba70010df79fab
(cherry picked from commit fd724409edb458be847b4d6e6f37ba41646a7e90)