Unpack/install all nodes before startup
Madan suggested that this might help vs. the pipelined
startup we were doing before, so let's try it.

We will unpack/install all nodes before we start up,
so that we can start the whole cluster at about the same
instant.

Note that this is probably closer to what we'd see in hardware.

The disadvantage is that we don't get a "clean" environment
with each invocation of start().

Change-Id: I0a45984f50e1ef25a7d8a836e46978fdbdff0bd0