Add type code into LispType enum to prevent arbitrary place holder
With the existing implementation, if we want to specify type value
8 to ECM Lisp type, we need to generate several place holders
inside LispType enumerator, otherwise ECM will be assigned a
subsequent number which is 5.
With this commit, we can specify arbitrary number for each Lisp
enum type, therefore, no need to add redundant placeholder for
the purpose of incrementing enum value.

Change-Id: I82a82a3f5bc313dc1f79aaa66a77889408e8c891