Enhance logging in NetconfSessionImpl
There are a few places where Throwable.getMessage() is emitted directly
instead of utilizing a Throwable argument to the Logger output methods.
Using debug as an example, when dealing with exception reporting cases,
it is, generally better to rely upon:

http://www.slf4j.org/api/org/slf4j/Logger.html#debug(java.lang.String, java.lang.Throwable)

than:

http://www.slf4j.org/api/org/slf4j/Logger.html#debug(java.lang.String)

This commit also cleans up the language of another log message.

Change-Id: I53a77286069c112aafc83ad8157a439a4d50e52e
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>