Build docker image from local source.
Currently we check out the code from gerrit to build the image, but this
makes it hard to specify a particular commit to build and impossible to
build a docker image from code that isn't yet checked in to git. This patch
will build the image from the local ONOS source.

Change-Id: I5936cb02d553aed3d9b6b5a260483759e36049c5