[ONOS-5853] FlowRuleBuilder: automatically set temp/perm field
Add a new method to the interface to allow setting the idle timeout
directly and it will automatically decide if the rule is permanent
or temporary.

Currently the rules created from OF switches would always be made
temporary, even if the idle-timeout was 0. This would cause ONOS
FlowRuleManager to believe they had expired and would remove them.

To reproduce set up a cluster and make a device change mastership, the
new master would have all rules as temporary, even the ones that were
permanent previously on the first master.

Change-Id: I6bca5ba4d0e2194efad21a98cbcd0ab040977f03