Ensure all Intents reaches installable state.
Previousely, there was an implicit assumption that
all Intent will reach installable state at same # of compilation steps.

ONOS-5855 opened a possibility for an Intent to be compiled down to
multiple kind of installables.
As a result, assumption that all Intents will reach installable
state after same # of compilation steps is likely to break.

This patch will wait for all compilation result to reach installable.

Change-Id: I307f39b660b586fc4cd98be9d1212c32163f0a90