[ONOS-6616] Revise VirtualPacketContext and DefaultVirtualPacketProvider
The constructor of VirtualPacketContext needs a parameter of DefaultVirtualPacketProvider type.
It is not flexible for us to use another packet provider to replace the default virtual packet provider.

To improve the code flexibility, I think it is better for us to use an interface type parameter in a method.

It alse seems redundant to use emit() method of DefaultVirtualPacketProvider in devirtualizeContext().
Thus, I think it will be more efficient to use core PacketService in VirtualPacketContext
when triger send() method.

Some other bugs are fixed.

Change-Id: I161a8929dc4e5a1d2ad716bc5da8b0b6f84340a9