Support for generic table IDs in flow rules
This change is needed to let appications specify tables in flow rules
using indentifiers that are not numeric, e.g. using table names as in a
P4 program.

This change deprecates the FlowRule.tableId() method, in favor of
FlowRule.table(), which returns an object of type TableId. However, it
does not deprecate the .forTable(int tableId) builder method, since it
is used extensivelly in the ONOS code base and from external
applications.

Deprecating FlowRule.tableId() does not seem to be such a painful move
since it is used less frequently, e.g. only 22 usages in the ONOS tree.

Change-Id: Idaaaab53d0c1e20a1f9c7de84de09ec3a42648b5