[CORD-2531] Missing flows and multicast groups
The rationale for this change is: unhandled exceptions
prevent Runnable to be scheduled in future.
We need to protect the body of runnable.

Related to [CORD-2532] and [CORD-2533].

In the pipeliner we use checkers and groups listener
to get notification about groups. The error scenario:
unhandled exception in the context of a group checker
will kill the checker. We create a multicast group, which
requires firstly a l2interfacegroup. The latter has been
already created in the past. No notification from groups
and no checker -> failure. Multicast group is not created
and forwarding objective in pending (forever).

Potentially, it can fix other errors scenario.

Change-Id: I6ea0548c112002b9ce415103891dc01431bc1dc8