Bug fix for refactored DestinationSet.
The issue was in the serialization. Since we enhanced DestinationSet
to use an Enumeration class, this class should also be registered
with KryoNamespace. Without registering the Enum class
DestinationSetType, the next hops were remaining local to the instance
which has inserted them into the Map and they did not propagate to
other instances.

Change-Id: I4e21481a5fe232892319232b8675cb083593f8f1