RestSBDevices with explicit proxy field
This patch allows to create RestSBDevices with explicit
isProxy (boolean) field.
This is important because previously, the RestDeviceProvider
did not allow this, resulting in new REST-based devices
being hidden behind other devices (thus not showing up in ONOS),
although a driver for these devices was available.

I also took the opportunity to slightly refactor the RestDeviceProvider.

Change-Id: I7c8ddf95ead357a35d54bc0bb6d36a2e4ca66f7a
Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>