Fixed that invalid links are created if lldp packets are received and there are no matched ports in onos.
Invalid links are created even if there are no matched ports in onos.
The root cause is that if lldp packets are received before ports are discovered, LinkDiscovery make a links using lldp packets's port name or port description.
In this case, a port number is generated using port name or port description. But this number is  not matched by the switch's port discovery driver.

So, this changes fix that do not create links if there are no ports in onos at LinkDiscovery.

ONOS-7852

Change-Id: Ib8477006d31fe39a770f2585475a9fedb759da76