Do not create a link if port is disabled.
There is some case regarding link creation.
Let's suppose that there are A and B devices on topology.

The link discovery drivers on onos will gather the link information from device(eg, switches).
In this case, if a port on the device A is unplugged or changed to down state but the link information(eg, LLDP) will be remained because TTL is not reached yet.

So, onos will re-create the link using link information that is gathered from link discovery driver.
This behavior result to link is re-created suddenly even if ports are disabled.

ONOS-7854

Change-Id: I224fb39fee92648600f9c9e40fcc97890d3ea9ef