Attempt to fix unit test
The execution time was originally set to a value that is too close to
the timeout. Therefore, there is chance that it will actually make it
in time. The purpose of this patch is to further delayed the execution
time and therefore get a better chance to create a true timeout.

Change-Id: Ib975db43b5f6d90cd91d0e44ee4cedcfd9c3cc6e