Prevent objective context from being executed multiple times
Consider the following sequence:
1. Flow objective issued
2. Cache timeout since the objective is not processed in time.
   onError()() will be triggered with ObjectiveError.INSTALLATIONTIMEOUT
3. While 2 is still being processed, the objective actually completed.
   onSuccess() will be triggered and therefore the same objective will be dequeued and executed again

The unit test was also wrong.
One objective context should only be executed once no matter the objective succeeded or failed.

In addition, we also fix unit test.

The execution time was originally set to a value that is too close to
the timeout. Therefore, there is chance that it will actually make it
in time. The purpose of this patch is to further delayed the execution
time and therefore get a better chance to create a true timeout.

Change-Id: I2501903bebee6fd9eefb03a71042c4a06ba4c42f