ONOS-7944 McastHandler of SegmentRouting considers only dual-homed sinks, not the sinks that are not tied with any host.
However, one of addSinks method of MulticastRouteService of mcast
application allows adding connect points each of which is a standalone
sink. In a scenario where all connect points are used as different sinks,
processSinksToBeAdded method in McastHandler service does not return
a true list of sinks to be added;

In overall,all connect points which are associated with HostId.NONE must
be added as different sinks. In the current situation, just one of them
is selected and added.

Change-Id: I3d4eb1ff865aca4f2fe85011a6a390374028a081