Fix PiGroupTranslatorImpl producing action profile members with ID 0
which is invalid in P4Runtime. This was observed when using action
profile group IDs with the last 2 bytes set to zeros, e.g. 0xec3b0000.
Now we generate the member ID as the hash of the group ID and the bucket
index, also adding a check for member IDs to be different than 0.

Change-Id: Id9fdf21d9704930703c3020b049cb19dafb1e590