[ONOS-8051] Device connected via BGP always shown available and reachable
In the command console, when "devices" command is executed, the call is routed
to DeviceManager. The DeviceManager queries the Provider of device periodically,
to check the isReachable status. If isReachable is true, isAvailable is eventually
set to true. Conversely, if isReachable is false, isAvailable is eventually set
to false.

Currently, BgptopologyProvider returns hard-coded "true" in isReachable() call.

This fix solves this by querying the current status from deviceService.
Also, since there is no "isReachable" call in deviceService, isAvailable() is used.

Change-Id: Ie6f7245c22232e0f867346f83b5aadaaf504031e