Support building with Bazel v1.0 on onos-1.15
Most of the changes are cherry-picked from the onos-2.2 branch. Requires
bumping grpc and protobuf to get compatible Bazel rules.

We provide preliminary support for building via the Bazel-provided
remote JDK. Unfortunately, the oldest remote JDK available is 9 which
doesn't work because of this bug with felix:
https://issues.apache.org/jira/browse/FELIX-5118

Fixing this issue would allow building with JDK 9 (or even 11), while
producing Java 8 compatible classes. This would make it easier for
developers swapping between onos-1.15 and master branches, as the
differences between Java versions would be handled by Bazel,
independently of the JDK installed in the host system. Until then, to
build onos-1.15, one needs to have JAVA_HOME pointing to a JDK 8.

Change-Id: I911f61e20ff983def44b73491aefc724a36fdcf9