Add a API to process a port description using consumer
Originally, there is a only API to discover the ports from device.
It fetches all ports from devices and returns as list.

If port fetching takes long, we cannot see the port update immediately until all ports are fetched.
So, I have added a API to fetch a port immediately, so that caller can get the updated port immediately when it is discovered.

Change-Id: I76571f9d015ee436edb8fe30ba5436352cc3c1ed