Prevents system workflow context from duplicated name assignment(with timestamp).
The name of the system workflow context is created based on the
msec timestamp('timestamp = System.currentTimeMillis'), so
duplicated system workflow context name can be assigned in a msec.

In order to prevent this duplicated name assignment,
It is changed to use UUID is used for guaranteeting
the uniqueness of system workflow context name instead of msec timestamp.

And name duplication checking is applied.

Change-Id: I4772e15d361af42bf61b6d47f93fee8a482814af